+++ BUILD
@@ -1394,7 +1394,6 @@ grpc_cc_library(
         "src/core/lib/compression/compression_args.cc",
         "src/core/lib/compression/compression_internal.cc",
         "src/core/lib/compression/message_compress.cc",
-        "src/core/lib/compression/null_compressor.cc",
         "src/core/lib/compression/stream_compression.cc",
         "src/core/lib/compression/stream_compression_gzip.cc",
         "src/core/lib/compression/stream_compression_identity.cc",
@@ -1562,7 +1561,6 @@ grpc_cc_library(
         "src/core/lib/compression/compression_args.h",
         "src/core/lib/compression/compression_internal.h",
         "src/core/lib/compression/message_compress.h",
-        "src/core/lib/compression/null_compressor.h",
         "src/core/lib/compression/stream_compression.h",
         "src/core/lib/compression/stream_compression_gzip.h",
         "src/core/lib/compression/stream_compression_identity.h",
