+++ gRPC-C++.podspec
@@ -528,7 +528,6 @@ Pod::Spec.new do |s|
                       'src/core/lib/compression/compression_args.h',
                       'src/core/lib/compression/compression_internal.h',
                       'src/core/lib/compression/message_compress.h',
-                      'src/core/lib/compression/null_compressor.h',
                       'src/core/lib/compression/stream_compression.h',
                       'src/core/lib/compression/stream_compression_gzip.h',
                       'src/core/lib/compression/stream_compression_identity.h',
@@ -1202,7 +1201,6 @@ Pod::Spec.new do |s|
                               'src/core/lib/compression/compression_args.h',
                               'src/core/lib/compression/compression_internal.h',
                               'src/core/lib/compression/message_compress.h',
-                              'src/core/lib/compression/null_compressor.h',
                               'src/core/lib/compression/stream_compression.h',
                               'src/core/lib/compression/stream_compression_gzip.h',
                               'src/core/lib/compression/stream_compression_identity.h',
