+++ gRPC-Core.podspec
@@ -858,8 +858,6 @@ Pod::Spec.new do |s|
                       'src/core/lib/compression/compression_internal.h',
                       'src/core/lib/compression/message_compress.cc',
                       'src/core/lib/compression/message_compress.h',
-                      'src/core/lib/compression/null_compressor.cc',
-                      'src/core/lib/compression/null_compressor.h',
                       'src/core/lib/compression/stream_compression.cc',
                       'src/core/lib/compression/stream_compression.h',
                       'src/core/lib/compression/stream_compression_gzip.cc',
@@ -1791,7 +1789,6 @@ Pod::Spec.new do |s|
                               'src/core/lib/compression/compression_args.h',
                               'src/core/lib/compression/compression_internal.h',
                               'src/core/lib/compression/message_compress.h',
-                              'src/core/lib/compression/null_compressor.h',
                               'src/core/lib/compression/stream_compression.h',
                               'src/core/lib/compression/stream_compression_gzip.h',
                               'src/core/lib/compression/stream_compression_identity.h',
