+++ grpc.gyp
@@ -813,7 +813,6 @@
         'src/core/lib/compression/compression_args.cc',
         'src/core/lib/compression/compression_internal.cc',
         'src/core/lib/compression/message_compress.cc',
-        'src/core/lib/compression/null_compressor.cc',
         'src/core/lib/compression/stream_compression.cc',
         'src/core/lib/compression/stream_compression_gzip.cc',
         'src/core/lib/compression/stream_compression_identity.cc',
@@ -1279,7 +1278,6 @@
         'src/core/lib/compression/compression_args.cc',
         'src/core/lib/compression/compression_internal.cc',
         'src/core/lib/compression/message_compress.cc',
-        'src/core/lib/compression/null_compressor.cc',
         'src/core/lib/compression/stream_compression.cc',
         'src/core/lib/compression/stream_compression_gzip.cc',
         'src/core/lib/compression/stream_compression_identity.cc',
