+++ src/core/lib/compression/message_compress.h
@@ -84,4 +84,18 @@ class CompressorRegistry {
   std::unordered_map<std::string, Compressor*> compressors;
 };
 
+/* A Compressor that does nothing */
+class NullCompressor : public Compressor {
+ public:
+  std::string encodingType();
+
+  void start();
+
+  void stop();
+
+  int msg_compress(grpc_slice_buffer* input, grpc_slice_buffer* output);
+
+  int msg_decompress(grpc_slice_buffer* input, grpc_slice_buffer* output);
+};
+
 #endif /* GRPC_CORE_LIB_COMPRESSION_MESSAGE_COMPRESS_H */
