--- src/core/lib/compression/null_compressor.h
@@ -1,34 +1,0 @@
-// Copyright 2021 gRPC authors.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef GRPC_CORE_LIB_COMPRESSION_NULL_COMPRESSOR_H
-#define GRPC_CORE_LIB_COMPRESSION_NULL_COMPRESSOR_H
-
-#include "src/core/lib/compression/message_compress.h"
-
-/* A Compressor that does nothing */
-class NullCompressor : public Compressor {
- public:
-  std::string encodingType();
-
-  void start();
-
-  void stop();
-
-  int msg_compress(grpc_slice_buffer* input, grpc_slice_buffer* output);
-
-  int msg_decompress(grpc_slice_buffer* input, grpc_slice_buffer* output);
-};
-
-#endif /* GRPC_CORE_LIB_COMPRESSION_NULL_COMPRESSOR_H */
