+++ src/python/grpcio/grpc_core_dependencies.py
@@ -354,7 +354,6 @@ CORE_SOURCE_FILES = [
     'src/core/lib/compression/compression_args.cc',
     'src/core/lib/compression/compression_internal.cc',
     'src/core/lib/compression/message_compress.cc',
-    'src/core/lib/compression/null_compressor.cc',
     'src/core/lib/compression/stream_compression.cc',
     'src/core/lib/compression/stream_compression_gzip.cc',
     'src/core/lib/compression/stream_compression_identity.cc',
