+++ tools/doxygen/Doxyfile.c++.internal
@@ -1703,8 +1703,6 @@ src/core/lib/compression/compression_internal.cc \
 src/core/lib/compression/compression_internal.h \
 src/core/lib/compression/message_compress.cc \
 src/core/lib/compression/message_compress.h \
-src/core/lib/compression/null_compressor.cc \
-src/core/lib/compression/null_compressor.h \
 src/core/lib/compression/stream_compression.cc \
 src/core/lib/compression/stream_compression.h \
 src/core/lib/compression/stream_compression_gzip.cc \
